refactor: delete useless GuiceTest and update package structure for UserServiceTest
